Crate google_sheets4

This documentation was generated from Sheets crate version 1.0.9+20190625, where 20190625 is the exact revision of the sheets:v4 schema built by the mako code generator v1.0.9.

Everything else about the Sheets v4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

spreadsheets

batch update, create, developer metadata get, developer metadata search, get, get by data filter, sheets copy to, values append, values batch clear, values batch clear by data filter, values batch get, values batch get by data filter, values batch update, values batch update by data filter, values clear, values get and values update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.spreadsheets().get_by_data_filter(...).doit()
let r = hub.spreadsheets().create(...).doit()
let r = hub.spreadsheets().values_batch_clear(...).doit()
let r = hub.spreadsheets().get(...).doit()
let r = hub.spreadsheets().values_batch_get(...).doit()
let r = hub.spreadsheets().values_append(...).doit()
let r = hub.spreadsheets().values_get(...).doit()
let r = hub.spreadsheets().sheets_copy_to(...).doit()
let r = hub.spreadsheets().values_clear(...).doit()
let r = hub.spreadsheets().values_update(...).doit()
let r = hub.spreadsheets().values_batch_update(...).doit()
let r = hub.spreadsheets().values_batch_get_by_data_filter(...).doit()
let r = hub.spreadsheets().values_batch_clear_by_data_filter(...).doit()
let r = hub.spreadsheets().values_batch_update_by_data_filter(...).doit()
let r = hub.spreadsheets().batch_update(...).doit()
let r = hub.spreadsheets().developer_metadata_search(...).doit()
let r = hub.spreadsheets().developer_metadata_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-sheets4 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_sheets4 as sheets4;
use sheets4::ValueRange;
use sheets4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sheets4::Sheets;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Sheets::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = ValueRange::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.spreadsheets().values_append(req, "spreadsheetId", "range")
             .value_input_option("justo")
             .response_value_render_option("amet.")
             .response_date_time_render_option("erat")
             .insert_data_option("labore")
             .include_values_in_response(true)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AddBandingRequest
Adds a new banded range to the spreadsheet.

AddBandingResponse
The result of adding a banded range.

AddChartRequest
Adds a chart to a sheet in the spreadsheet.

AddChartResponse
The result of adding a chart to a spreadsheet.

AddConditionalFormatRuleRequest
Adds a new conditional format rule at the given index. All subsequent rules' indexes are incremented.

AddDimensionGroupRequest
Creates a group over the specified range.

AddDimensionGroupResponse
The result of adding a group.

AddFilterViewRequest
Adds a filter view.

AddFilterViewResponse
The result of adding a filter view.

AddNamedRangeRequest
Adds a named range to the spreadsheet.

AddNamedRangeResponse
The result of adding a named range.

AddProtectedRangeRequest
Adds a new protected range.

AddProtectedRangeResponse
The result of adding a new protected range.

AddSheetRequest
Adds a new sheet. When a sheet is added at a given index, all subsequent sheets' indexes are incremented. To add an object sheet, use AddChartRequest instead and specify EmbeddedObjectPosition.sheetId or EmbeddedObjectPosition.newSheet.

AddSheetResponse
The result of adding a sheet.

AppendCellsRequest
Adds new cells after the last row with data in a sheet, inserting new rows into the sheet if necessary.

AppendDimensionRequest
Appends rows or columns to the end of a sheet.

AppendValuesResponse
The response when updating a range of values in a spreadsheet.

AutoFillRequest
Fills in more data based on existing data.

AutoResizeDimensionsRequest
Automatically resizes one or more dimensions based on the contents of the cells in that dimension.

BandedRange
A banded (alternating colors) range in a sheet.

BandingProperties
Properties referring a single dimension (either row or column). If both BandedRange.row_properties and BandedRange.column_properties are set, the fill colors are applied to cells according to the following rules:

BasicChartAxis
An axis of the chart. A chart may not have more than one axis per axis position.

BasicChartDomain
The domain of a chart. For example, if charting stock prices over time, this would be the date.

BasicChartSeries
A single series of data in a chart. For example, if charting stock prices over time, multiple series may exist, one for the "Open Price", "High Price", "Low Price" and "Close Price".

BasicChartSpec
The specification for a basic chart. See BasicChartType for the list of charts this supports.

BasicFilter
The default filter associated with a sheet.

BatchClearValuesByDataFilterRequest
The request for clearing more than one range selected by a DataFilter in a spreadsheet.

BatchClearValuesByDataFilterResponse
The response when clearing a range of values selected with DataFilters in a spreadsheet.

BatchClearValuesRequest
The request for clearing more than one range of values in a spreadsheet.

BatchClearValuesResponse
The response when clearing a range of values in a spreadsheet.

BatchGetValuesByDataFilterRequest
The request for retrieving a range of values in a spreadsheet selected by a set of DataFilters.

BatchGetValuesByDataFilterResponse
The response when retrieving more than one range of values in a spreadsheet selected by DataFilters.

BatchGetValuesResponse
The response when retrieving more than one range of values in a spreadsheet.

BatchUpdateSpreadsheetRequest
The request for updating any aspect of a spreadsheet.

BatchUpdateSpreadsheetResponse
The reply for batch updating a spreadsheet.

BatchUpdateValuesByDataFilterRequest
The request for updating more than one range of values in a spreadsheet.

BatchUpdateValuesByDataFilterResponse
The response when updating a range of values in a spreadsheet.

BatchUpdateValuesRequest
The request for updating more than one range of values in a spreadsheet.

BatchUpdateValuesResponse
The response when updating a range of values in a spreadsheet.

BooleanCondition
A condition that can evaluate to true or false. BooleanConditions are used by conditional formatting, data validation, and the criteria in filters.

BooleanRule
A rule that may or may not match, depending on the condition.

Border
A border along a cell.

Borders
The borders of the cell.

BubbleChartSpec
A bubble chart.

CandlestickChartSpec
A candlestick chart.

CandlestickData
The Candlestick chart data, each containing the low, open, close, and high values for a series.

CandlestickDomain
The domain of a CandlestickChart.

CandlestickSeries
The series of a CandlestickData.

CellData
Data about a specific cell.

CellFormat
The format of a cell.

ChartData
The data included in a domain or series.

ChartSourceRange
Source ranges for a chart.

ChartSpec
The specifications of a chart.

Chunk

ClearBasicFilterRequest
Clears the basic filter, if any exists on the sheet.

ClearValuesRequest
The request for clearing a range of values in a spreadsheet.

ClearValuesResponse
The response when clearing a range of values in a spreadsheet.

Color
Represents a color in the RGBA color space. This representation is designed for simplicity of conversion to/from color representations in various languages over compactness; for example, the fields of this representation can be trivially provided to the constructor of "java.awt.Color" in Java; it can also be trivially provided to UIColor's "+colorWithRed:green:blue:alpha" method in iOS; and, with just a little work, it can be easily formatted into a CSS "rgba()" string in JavaScript, as well.

ConditionValue
The value of the condition.

ConditionalFormatRule
A rule describing a conditional format.

ContentRange
Implements the Content-Range header, for serialization only

CopyPasteRequest
Copies data from the source to the destination.

CopySheetToAnotherSpreadsheetRequest
The request to copy a sheet across spreadsheets.

CreateDeveloperMetadataRequest
A request to create developer metadata.

CreateDeveloperMetadataResponse
The response from creating developer metadata.

CutPasteRequest
Moves data from the source to the destination.

DataFilter
Filter that describes what data should be selected or returned from a request.

DataFilterValueRange
A range of values whose location is specified by a DataFilter.

DataValidationRule
A data validation rule.

DateTimeRule
Allows you to organize the date-time values in a source data column into buckets based on selected parts of their date or time values. For example, consider a pivot table showing sales transactions by date:

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeleteBandingRequest
Removes the banded range with the given ID from the spreadsheet.

DeleteConditionalFormatRuleRequest
Deletes a conditional format rule at the given index. All subsequent rules' indexes are decremented.

DeleteConditionalFormatRuleResponse
The result of deleting a conditional format rule.

DeleteDeveloperMetadataRequest
A request to delete developer metadata.

DeleteDeveloperMetadataResponse
The response from deleting developer metadata.

DeleteDimensionGroupRequest
Deletes a group over the specified range by decrementing the depth of the dimensions in the range.

DeleteDimensionGroupResponse
The result of deleting a group.

DeleteDimensionRequest
Deletes the dimensions from the sheet.

DeleteEmbeddedObjectRequest
Deletes the embedded object with the given ID.

DeleteFilterViewRequest
Deletes a particular filter view.

DeleteNamedRangeRequest
Removes the named range with the given ID from the spreadsheet.

DeleteProtectedRangeRequest
Deletes the protected range with the given ID.

DeleteRangeRequest
Deletes a range of cells, shifting other cells into the deleted area.

DeleteSheetRequest
Deletes the requested sheet.

DeveloperMetadata
Developer metadata associated with a location or object in a spreadsheet. Developer metadata may be used to associate arbitrary data with various parts of a spreadsheet and will remain associated at those locations as they move around and the spreadsheet is edited. For example, if developer metadata is associated with row 5 and another row is then subsequently inserted above row 5, that original metadata will still be associated with the row it was first associated with (what is now row 6). If the associated object is deleted its metadata is deleted too.

DeveloperMetadataLocation
A location where metadata may be associated in a spreadsheet.

DeveloperMetadataLookup
Selects DeveloperMetadata that matches all of the specified fields. For example, if only a metadata ID is specified this considers the DeveloperMetadata with that particular unique ID. If a metadata key is specified, this considers all developer metadata with that key. If a key, visibility, and location type are all specified, this considers all developer metadata with that key and visibility that are associated with a location of that type. In general, this selects all DeveloperMetadata that matches the intersection of all the specified fields; any field or combination of fields may be specified.

DimensionGroup
A group over an interval of rows or columns on a sheet, which can contain or be contained within other groups. A group can be collapsed or expanded as a unit on the sheet.

DimensionProperties
Properties about a dimension.

DimensionRange
A range along a single dimension on a sheet. All indexes are zero-based. Indexes are half open: the start index is inclusive and the end index is exclusive. Missing indexes indicate the range is unbounded on that side.

DummyNetworkStream

DuplicateFilterViewRequest
Duplicates a particular filter view.

DuplicateFilterViewResponse
The result of a filter view being duplicated.

DuplicateSheetRequest
Duplicates the contents of a sheet.

DuplicateSheetResponse
The result of duplicating a sheet.

Editors
The editors of a protected range.

EmbeddedChart
A chart embedded in a sheet.

EmbeddedObjectPosition
The position of an embedded object such as a chart.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ErrorValue
An error in a cell.

ExtendedValue
The kinds of value that a cell in a spreadsheet can have.

FilterCriteria
Criteria for showing/hiding rows in a filter or filter view.

FilterView
A filter view.

FindReplaceRequest
Finds and replaces data in cells over a range, sheet, or all sheets.

FindReplaceResponse
The result of the find/replace.

GetSpreadsheetByDataFilterRequest
The request for retrieving a Spreadsheet.

GradientRule
A rule that applies a gradient color scale format, based on the interpolation points listed. The format of a cell will vary based on its contents as compared to the values of the interpolation points.

GridCoordinate
A coordinate in a sheet. All indexes are zero-based.

GridData
Data in the grid, as well as metadata about the dimensions.

GridProperties
Properties of a grid.

GridRange
A range on a sheet. All indexes are zero-based. Indexes are half open, e.g the start index is inclusive and the end index is exclusive -- [start_index, end_index). Missing indexes indicate the range is unbounded on that side.

HistogramChartSpec
A histogram chart. A histogram chart groups data items into bins, displaying each bin as a column of stacked items. Histograms are used to display the distribution of a dataset. Each column of items represents a range into which those items fall. The number of bins can be chosen automatically or specified explicitly.

HistogramRule
Allows you to organize the numeric values in a source data column into buckets of a constant size. All values from HistogramRule.start to HistogramRule.end are placed into groups of size HistogramRule.interval. In addition, all values below HistogramRule.start are placed in one group, and all values above HistogramRule.end are placed in another. Only HistogramRule.interval is required, though if HistogramRule.start and HistogramRule.end are both provided, HistogramRule.start must be less than HistogramRule.end. For example, a pivot table showing average purchase amount by age that has 50+ rows:

HistogramSeries
A histogram series containing the series color and data.

InsertDimensionRequest
Inserts rows or columns in a sheet at a particular index.

InsertRangeRequest
Inserts cells into a range, shifting the existing cells over or down.

InterpolationPoint
A single interpolation point on a gradient conditional format. These pin the gradient color scale according to the color, type and value chosen.

IterativeCalculationSettings
Settings to control how circular dependencies are resolved with iterative calculation.

JsonServerError
A utility type which can decode a server response that indicates error

LineStyle
Properties that describe the style of a line.

ManualRule
Allows you to manually organize the values in a source data column into buckets with names of your choosing. For example, a pivot table that aggregates population by state:

ManualRuleGroup
A group name and a list of items from the source data that should be placed in the group with this name.

MatchedDeveloperMetadata
A developer metadata entry and the data filters specified in the original request that matched it.

MatchedValueRange
A value range that was matched by one or more data filers.

MergeCellsRequest
Merges all cells in the range.

MethodInfo
Contains information about an API request.

MoveDimensionRequest
Moves one or more rows or columns.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NamedRange
A named range.

NumberFormat
The number format of a cell.

OrgChartSpec
An org chart. Org charts require a unique set of labels in labels and may optionally include parent_labels and tooltips. parent_labels contain, for each node, the label identifying the parent node. tooltips contain, for each node, an optional tooltip.

OverlayPosition
The location an object is overlaid on top of a grid.

Padding
The amount of padding around the cell, in pixels. When updating padding, every field must be specified.

PasteDataRequest
Inserts data into the spreadsheet starting at the specified coordinate.

PieChartSpec
A pie chart.

PivotFilterCriteria
Criteria for showing/hiding rows in a pivot table.

PivotGroup
A single grouping (either row or column) in a pivot table.

PivotGroupRule
An optional setting on a PivotGroup that defines buckets for the values in the source data column rather than breaking out each individual value. Only one PivotGroup with a group rule may be added for each column in the source data, though on any given column you may add both a PivotGroup that has a rule and a PivotGroup that does not.

PivotGroupSortValueBucket
Information about which values in a pivot group should be used for sorting.

PivotGroupValueMetadata
Metadata about a value in a pivot grouping.

PivotTable
A pivot table.

PivotValue
The definition of how a value in a pivot table should be calculated.

ProtectedRange
A protected range.

RandomizeRangeRequest
Randomizes the order of the rows in a range.

RangeResponseHeader

RepeatCellRequest
Updates all cells in the range to the values in the given Cell object. Only the fields listed in the fields field are updated; others are unchanged.

Request
A single kind of update to apply to a spreadsheet.

Response
A single response from an update.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

RowData
Data about each cell in a row.

SearchDeveloperMetadataRequest
A request to retrieve all developer metadata matching the set of specified criteria.

SearchDeveloperMetadataResponse
A reply to a developer metadata search request.

ServerError

ServerMessage

SetBasicFilterRequest
Sets the basic filter associated with a sheet.

SetDataValidationRequest
Sets a data validation rule to every cell in the range. To clear validation in a range, call this with no rule specified.

Sheet
A sheet in a spreadsheet.

SheetProperties
Properties of a sheet.

Sheets
Central instance to access all Sheets related resource activities

SortRangeRequest
Sorts data in rows based on a sort order per column.

SortSpec
A sort order associated with a specific column or row.

SourceAndDestination
A combination of a source range and how to extend that source.

Spreadsheet
Resource that represents a spreadsheet.

SpreadsheetBatchUpdateCall
Applies one or more updates to the spreadsheet.

SpreadsheetCreateCall
Creates a spreadsheet, returning the newly created spreadsheet.

SpreadsheetDeveloperMetadataGetCall
Returns the developer metadata with the specified ID. The caller must specify the spreadsheet ID and the developer metadata's unique metadataId.

SpreadsheetDeveloperMetadataSearchCall
Returns all developer metadata matching the specified DataFilter. If the provided DataFilter represents a DeveloperMetadataLookup object, this will return all DeveloperMetadata entries selected by it. If the DataFilter represents a location in a spreadsheet, this will return all developer metadata associated with locations intersecting that region.

SpreadsheetGetByDataFilterCall
Returns the spreadsheet at the given ID. The caller must specify the spreadsheet ID.

SpreadsheetGetCall
Returns the spreadsheet at the given ID. The caller must specify the spreadsheet ID.

SpreadsheetMethods
A builder providing access to all methods supported on spreadsheet resources. It is not used directly, but through the Sheets hub.

SpreadsheetProperties
Properties of a spreadsheet.

SpreadsheetSheetCopyToCall
Copies a single sheet from a spreadsheet to another spreadsheet. Returns the properties of the newly created sheet.

SpreadsheetValueAppendCall
Appends values to a spreadsheet. The input range is used to search for existing data and find a "table" within that range. Values will be appended to the next row of the table, starting with the first column of the table. See the guide and sample code for specific details of how tables are detected and data is appended.

SpreadsheetValueBatchClearByDataFilterCall
Clears one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more DataFilters. Ranges matching any of the specified data filters will be cleared. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueBatchClearCall
Clears one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more ranges. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueBatchGetByDataFilterCall
Returns one or more ranges of values that match the specified data filters. The caller must specify the spreadsheet ID and one or more DataFilters. Ranges that match any of the data filters in the request will be returned.

SpreadsheetValueBatchGetCall
Returns one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more ranges.

SpreadsheetValueBatchUpdateByDataFilterCall
Sets values in one or more ranges of a spreadsheet. The caller must specify the spreadsheet ID, a valueInputOption, and one or more DataFilterValueRanges.

SpreadsheetValueBatchUpdateCall
Sets values in one or more ranges of a spreadsheet. The caller must specify the spreadsheet ID, a valueInputOption, and one or more ValueRanges.

SpreadsheetValueClearCall
Clears values from a spreadsheet. The caller must specify the spreadsheet ID and range. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueGetCall
Returns a range of values from a spreadsheet. The caller must specify the spreadsheet ID and a range.

SpreadsheetValueUpdateCall
Sets values in a range of a spreadsheet. The caller must specify the spreadsheet ID, range, and a valueInputOption.

TextFormat
The format of a run of text in a cell. Absent values indicate that the field isn't specified.

TextFormatRun
A run of a text format. The format of this run continues until the start index of the next run. When updating, all fields must be set.

TextPosition
Position settings for text.

TextRotation
The rotation applied to text in a cell.

TextToColumnsRequest
Splits a column of text into multiple columns, based on a delimiter in each cell.

TreemapChartColorScale
A color scale for a treemap chart.

TreemapChartSpec
A Treemap chart.

UnmergeCellsRequest
Unmerges cells in the given range.

UpdateBandingRequest
Updates properties of the supplied banded range.

UpdateBordersRequest
Updates the borders of a range. If a field is not set in the request, that means the border remains as-is. For example, with two subsequent UpdateBordersRequest:

UpdateCellsRequest
Updates all cells in a range with new data.

UpdateChartSpecRequest
Updates a chart's specifications. (This does not move or resize a chart. To move or resize a chart, use UpdateEmbeddedObjectPositionRequest.)

UpdateConditionalFormatRuleRequest
Updates a conditional format rule at the given index, or moves a conditional format rule to another index.

UpdateConditionalFormatRuleResponse
The result of updating a conditional format rule.

UpdateDeveloperMetadataRequest
A request to update properties of developer metadata. Updates the properties of the developer metadata selected by the filters to the values provided in the DeveloperMetadata resource. Callers must specify the properties they wish to update in the fields parameter, as well as specify at least one DataFilter matching the metadata they wish to update.

UpdateDeveloperMetadataResponse
The response from updating developer metadata.

UpdateDimensionGroupRequest
Updates the state of the specified group.

UpdateDimensionPropertiesRequest
Updates properties of dimensions within the specified range.

UpdateEmbeddedObjectPositionRequest
Update an embedded object's position (such as a moving or resizing a chart or image).

UpdateEmbeddedObjectPositionResponse
The result of updating an embedded object's position.

UpdateFilterViewRequest
Updates properties of the filter view.

UpdateNamedRangeRequest
Updates properties of the named range with the specified namedRangeId.

UpdateProtectedRangeRequest
Updates an existing protected range with the specified protectedRangeId.

UpdateSheetPropertiesRequest
Updates properties of the sheet with the specified sheetId.

UpdateSpreadsheetPropertiesRequest
Updates properties of a spreadsheet.

UpdateValuesByDataFilterResponse
The response when updating a range of values by a data filter in a spreadsheet.

UpdateValuesResponse
The response when updating a range of values in a spreadsheet.

ValueRange
Data within a range of the spreadsheet.

WaterfallChartColumnStyle
Styles for a waterfall chart column.

WaterfallChartCustomSubtotal
A custom subtotal column for a waterfall chart series.

WaterfallChartDomain
The domain of a waterfall chart.

WaterfallChartSeries
A single series of data for a waterfall chart.

WaterfallChartSpec
A waterfall chart.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.