Module google_sheets4::api

Structs

AddBandingRequest
Adds a new banded range to the spreadsheet.

AddBandingResponse
The result of adding a banded range.

AddChartRequest
Adds a chart to a sheet in the spreadsheet.

AddChartResponse
The result of adding a chart to a spreadsheet.

AddConditionalFormatRuleRequest
Adds a new conditional format rule at the given index. All subsequent rules' indexes are incremented.

AddDataSourceRequest
Adds a data source. After the data source is added successfully, an associated DATA_SOURCE sheet is created and an execution is triggered to refresh the sheet to read data from the data source. The request requires an additional bigquery.readonly OAuth scope.

AddDataSourceResponse
The result of adding a data source.

AddDimensionGroupRequest
Creates a group over the specified range. If the requested range is a superset of the range of an existing group G, then the depth of G is incremented and this new group G' has the depth of that group. For example, a group [C:D, depth 1] + [B:E] results in groups [B:E, depth 1] and [C:D, depth 2]. If the requested range is a subset of the range of an existing group G, then the depth of the new group G' becomes one greater than the depth of G. For example, a group [B:E, depth 1] + [C:D] results in groups [B:E, depth 1] and [C:D, depth 2]. If the requested range starts before and ends within, or starts within and ends after, the range of an existing group G, then the range of the existing group G becomes the union of the ranges, and the new group G' has depth one greater than the depth of G and range as the intersection of the ranges. For example, a group [B:D, depth 1] + [C:E] results in groups [B:E, depth 1] and [C:D, depth 2].

AddDimensionGroupResponse
The result of adding a group.

AddFilterViewRequest
Adds a filter view.

AddFilterViewResponse
The result of adding a filter view.

AddNamedRangeRequest
Adds a named range to the spreadsheet.

AddNamedRangeResponse
The result of adding a named range.

AddProtectedRangeRequest
Adds a new protected range.

AddProtectedRangeResponse
The result of adding a new protected range.

AddSheetRequest
Adds a new sheet. When a sheet is added at a given index, all subsequent sheets' indexes are incremented. To add an object sheet, use AddChartRequest instead and specify EmbeddedObjectPosition.sheetId or EmbeddedObjectPosition.newSheet.

AddSheetResponse
The result of adding a sheet.

AddSlicerRequest
Adds a slicer to a sheet in the spreadsheet.

AddSlicerResponse
The result of adding a slicer to a spreadsheet.

AppendCellsRequest
Adds new cells after the last row with data in a sheet, inserting new rows into the sheet if necessary.

AppendDimensionRequest
Appends rows or columns to the end of a sheet.

AppendValuesResponse
The response when updating a range of values in a spreadsheet.

AutoFillRequest
Fills in more data based on existing data.

AutoResizeDimensionsRequest
Automatically resizes one or more dimensions based on the contents of the cells in that dimension.

BandedRange
A banded (alternating colors) range in a sheet.

BandingProperties
Properties referring a single dimension (either row or column). If both BandedRange.row_properties and BandedRange.column_properties are set, the fill colors are applied to cells according to the following rules: * header_color and footer_color take priority over band colors. * first_band_color takes priority over second_band_color. * row_properties takes priority over column_properties. For example, the first row color takes priority over the first column color, but the first column color takes priority over the second row color. Similarly, the row header takes priority over the column header in the top left cell, but the column header takes priority over the first row color if the row header is not set.

BaselineValueFormat
Formatting options for baseline value.

BasicChartAxis
An axis of the chart. A chart may not have more than one axis per axis position.

BasicChartDomain
The domain of a chart. For example, if charting stock prices over time, this would be the date.

BasicChartSeries
A single series of data in a chart. For example, if charting stock prices over time, multiple series may exist, one for the "Open Price", "High Price", "Low Price" and "Close Price".

BasicChartSpec
The specification for a basic chart. See BasicChartType for the list of charts this supports.

BasicFilter
The default filter associated with a sheet.

BasicSeriesDataPointStyleOverride
Style override settings for a single series data point.

BatchClearValuesByDataFilterRequest
The request for clearing more than one range selected by a DataFilter in a spreadsheet.

BatchClearValuesByDataFilterResponse
The response when clearing a range of values selected with DataFilters in a spreadsheet.

BatchClearValuesRequest
The request for clearing more than one range of values in a spreadsheet.

BatchClearValuesResponse
The response when clearing a range of values in a spreadsheet.

BatchGetValuesByDataFilterRequest
The request for retrieving a range of values in a spreadsheet selected by a set of DataFilters.

BatchGetValuesByDataFilterResponse
The response when retrieving more than one range of values in a spreadsheet selected by DataFilters.

BatchGetValuesResponse
The response when retrieving more than one range of values in a spreadsheet.

BatchUpdateSpreadsheetRequest
The request for updating any aspect of a spreadsheet.

BatchUpdateSpreadsheetResponse
The reply for batch updating a spreadsheet.

BatchUpdateValuesByDataFilterRequest
The request for updating more than one range of values in a spreadsheet.

BatchUpdateValuesByDataFilterResponse
The response when updating a range of values in a spreadsheet.

BatchUpdateValuesRequest
The request for updating more than one range of values in a spreadsheet.

BatchUpdateValuesResponse
The response when updating a range of values in a spreadsheet.

BigQueryDataSourceSpec
The specification of a BigQuery data source that's connected to a sheet.

BigQueryQuerySpec
Specifies a custom BigQuery query.

BigQueryTableSpec
Specifies a BigQuery table definition. Only native tables is allowed.

BooleanCondition
A condition that can evaluate to true or false. BooleanConditions are used by conditional formatting, data validation, and the criteria in filters.

BooleanRule
A rule that may or may not match, depending on the condition.

Border
A border along a cell.

Borders
The borders of the cell.

BubbleChartSpec
A bubble chart.

CandlestickChartSpec
A candlestick chart.

CandlestickData
The Candlestick chart data, each containing the low, open, close, and high values for a series.

CandlestickDomain
The domain of a CandlestickChart.

CandlestickSeries
The series of a CandlestickData.

CellData
Data about a specific cell.

CellFormat
The format of a cell.

ChartAxisViewWindowOptions
The options that define a "view window" for a chart (such as the visible values in an axis).

ChartCustomNumberFormatOptions
Custom number formatting options for chart attributes.

ChartData
The data included in a domain or series.

ChartDateTimeRule
Allows you to organize the date-time values in a source data column into buckets based on selected parts of their date or time values.

ChartGroupRule
An optional setting on the ChartData of the domain of a data source chart that defines buckets for the values in the domain rather than breaking out each individual value. For example, when plotting a data source chart, you can specify a histogram rule on the domain (it should only contain numeric values), grouping its values into buckets. Any values of a chart series that fall into the same bucket are aggregated based on the aggregate_type.

ChartHistogramRule
Allows you to organize numeric values in a source data column into buckets of constant size.

ChartSourceRange
Source ranges for a chart.

ChartSpec
The specifications of a chart.

ClearBasicFilterRequest
Clears the basic filter, if any exists on the sheet.

ClearValuesRequest
The request for clearing a range of values in a spreadsheet.

ClearValuesResponse
The response when clearing a range of values in a spreadsheet.

Color
Represents a color in the RGBA color space. This representation is designed for simplicity of conversion to/from color representations in various languages over compactness. For example, the fields of this representation can be trivially provided to the constructor of java.awt.Color in Java; it can also be trivially provided to UIColor's +colorWithRed:green:blue:alpha method in iOS; and, with just a little work, it can be easily formatted into a CSS rgba() string in JavaScript. This reference page doesn't carry information about the absolute color space that should be used to interpret the RGB value (e.g. sRGB, Adobe RGB, DCI-P3, BT.2020, etc.). By default, applications should assume the sRGB color space. When color equality needs to be decided, implementations, unless documented otherwise, treat two colors as equal if all their red, green, blue, and alpha values each differ by at most 1e-5. Example (Java): import com.google.type.Color; // ... public static java.awt.Color fromProto(Color protocolor) { float alpha = protocolor.hasAlpha() ? protocolor.getAlpha().getValue() : 1.0; return new java.awt.Color( protocolor.getRed(), protocolor.getGreen(), protocolor.getBlue(), alpha); } public static Color toProto(java.awt.Color color) { float red = (float) color.getRed(); float green = (float) color.getGreen(); float blue = (float) color.getBlue(); float denominator = 255.0; Color.Builder resultBuilder = Color .newBuilder() .setRed(red / denominator) .setGreen(green / denominator) .setBlue(blue / denominator); int alpha = color.getAlpha(); if (alpha != 255) { result.setAlpha( FloatValue .newBuilder() .setValue(((float) alpha) / denominator) .build()); } return resultBuilder.build(); } // ... Example (iOS / Obj-C): // ... static UIColor* fromProto(Color* protocolor) { float red = [protocolor red]; float green = [protocolor green]; float blue = [protocolor blue]; FloatValue* alpha_wrapper = [protocolor alpha]; float alpha = 1.0; if (alpha_wrapper != nil) { alpha = [alpha_wrapper value]; } return [UIColor colorWithRed:red green:green blue:blue alpha:alpha]; } static Color* toProto(UIColor* color) { CGFloat red, green, blue, alpha; if (![color getRed:&red green:&green blue:&blue alpha:&alpha]) { return nil; } Color* result = [[Color alloc] init]; [result setRed:red]; [result setGreen:green]; [result setBlue:blue]; if (alpha <= 0.9999) { [result setAlpha:floatWrapperWithValue(alpha)]; } [result autorelease]; return result; } // ... Example (JavaScript): // ... var protoToCssColor = function(rgb_color) { var redFrac = rgb_color.red || 0.0; var greenFrac = rgb_color.green || 0.0; var blueFrac = rgb_color.blue || 0.0; var red = Math.floor(redFrac * 255); var green = Math.floor(greenFrac * 255); var blue = Math.floor(blueFrac * 255); if (!('alpha' in rgb_color)) { return rgbToCssColor(red, green, blue); } var alphaFrac = rgb_color.alpha.value || 0.0; var rgbParams = [red, green, blue].join(','); return ['rgba(', rgbParams, ',', alphaFrac, ')'].join(''); }; var rgbToCssColor = function(red, green, blue) { var rgbNumber = new Number((red << 16) | (green << 8) | blue); var hexString = rgbNumber.toString(16); var missingZeros = 6 - hexString.length; var resultBuilder = ['#']; for (var i = 0; i < missingZeros; i++) { resultBuilder.push('0'); } resultBuilder.push(hexString); return resultBuilder.join(''); }; // ...

ColorStyle
A color value.

ConditionValue
The value of the condition.

ConditionalFormatRule
A rule describing a conditional format.

CopyPasteRequest
Copies data from the source to the destination.

CopySheetToAnotherSpreadsheetRequest
The request to copy a sheet across spreadsheets.

CreateDeveloperMetadataRequest
A request to create developer metadata.

CreateDeveloperMetadataResponse
The response from creating developer metadata.

CutPasteRequest
Moves data from the source to the destination.

DataExecutionStatus
The data execution status. A data execution is created to sync a data source object with the latest data from a DataSource. It is usually scheduled to run at background, you can check its state to tell if an execution completes There are several scenarios where a data execution is triggered to run: * Adding a data source creates an associated data source sheet as well as a data execution to sync the data from the data source to the sheet. * Updating a data source creates a data execution to refresh the associated data source sheet similarly. * You can send refresh request to explicitly refresh one or multiple data source objects.

DataFilter
Filter that describes what data should be selected or returned from a request.

DataFilterValueRange
A range of values whose location is specified by a DataFilter.

DataLabel
Settings for one set of data labels. Data labels are annotations that appear next to a set of data, such as the points on a line chart, and provide additional information about what the data represents, such as a text representation of the value behind that point on the graph.

DataSource
Information about an external data source in the spreadsheet.

DataSourceChartProperties
Properties of a data source chart.

DataSourceColumn
A column in a data source.

DataSourceColumnReference
An unique identifier that references a data source column.

DataSourceFormula
A data source formula.

DataSourceObjectReference
Reference to a data source object.

DataSourceObjectReferences
A list of references to data source objects.

DataSourceParameter
A parameter in a data source's query. The parameter allows the user to pass in values from the spreadsheet into a query.

DataSourceRefreshDailySchedule
A schedule for data to refresh every day in a given time interval.

DataSourceRefreshMonthlySchedule
A monthly schedule for data to refresh on specific days in the month in a given time interval.

DataSourceRefreshSchedule
Schedule for refreshing the data source. Data sources in the spreadsheet are refreshed within a time interval. You can specify the start time by clicking the Scheduled Refresh button in the Sheets editor, but the interval is fixed at 4 hours. For example, if you specify a start time of 8am , the refresh will take place between 8am and 12pm every day.

DataSourceRefreshWeeklySchedule
A weekly schedule for data to refresh on specific days in a given time interval.

DataSourceSheetDimensionRange
A range along a single dimension on a DATA_SOURCE sheet.

DataSourceSheetProperties
Additional properties of a DATA_SOURCE sheet.

DataSourceSpec
This specifies the details of the data source. For example, for BigQuery, this specifies information about the BigQuery source.

DataSourceTable
A data source table, which allows the user to import a static table of data from the DataSource into Sheets. This is also known as "Extract" in the Sheets editor.

DataValidationRule
A data validation rule.

DateTimeRule
Allows you to organize the date-time values in a source data column into buckets based on selected parts of their date or time values. For example, consider a pivot table showing sales transactions by date: +----------+--------------+ | Date | SUM of Sales | +----------+--------------+ | 1/1/2017 | $621.14 | | 2/3/2017 | $708.84 | | 5/8/2017 | $326.84 | ... +----------+--------------+ Applying a date-time group rule with a DateTimeRuleType of YEAR_MONTH results in the following pivot table. +--------------+--------------+ | Grouped Date | SUM of Sales | +--------------+--------------+ | 2017-Jan | $53,731.78 | | 2017-Feb | $83,475.32 | | 2017-Mar | $94,385.05 | ... +--------------+--------------+

DeleteBandingRequest
Removes the banded range with the given ID from the spreadsheet.

DeleteConditionalFormatRuleRequest
Deletes a conditional format rule at the given index. All subsequent rules' indexes are decremented.

DeleteConditionalFormatRuleResponse
The result of deleting a conditional format rule.

DeleteDataSourceRequest
Deletes a data source. The request also deletes the associated data source sheet, and unlinks all associated data source objects.

DeleteDeveloperMetadataRequest
A request to delete developer metadata.

DeleteDeveloperMetadataResponse
The response from deleting developer metadata.

DeleteDimensionGroupRequest
Deletes a group over the specified range by decrementing the depth of the dimensions in the range. For example, assume the sheet has a depth-1 group over B:E and a depth-2 group over C:D. Deleting a group over D:E leaves the sheet with a depth-1 group over B:D and a depth-2 group over C:C.

DeleteDimensionGroupResponse
The result of deleting a group.

DeleteDimensionRequest
Deletes the dimensions from the sheet.

DeleteDuplicatesRequest
Removes rows within this range that contain values in the specified columns that are duplicates of values in any previous row. Rows with identical values but different letter cases, formatting, or formulas are considered to be duplicates. This request also removes duplicate rows hidden from view (for example, due to a filter). When removing duplicates, the first instance of each duplicate row scanning from the top downwards is kept in the resulting range. Content outside of the specified range isn't removed, and rows considered duplicates do not have to be adjacent to each other in the range.

DeleteDuplicatesResponse
The result of removing duplicates in a range.

DeleteEmbeddedObjectRequest
Deletes the embedded object with the given ID.

DeleteFilterViewRequest
Deletes a particular filter view.

DeleteNamedRangeRequest
Removes the named range with the given ID from the spreadsheet.

DeleteProtectedRangeRequest
Deletes the protected range with the given ID.

DeleteRangeRequest
Deletes a range of cells, shifting other cells into the deleted area.

DeleteSheetRequest
Deletes the requested sheet.

DeveloperMetadata
Developer metadata associated with a location or object in a spreadsheet. Developer metadata may be used to associate arbitrary data with various parts of a spreadsheet and will remain associated at those locations as they move around and the spreadsheet is edited. For example, if developer metadata is associated with row 5 and another row is then subsequently inserted above row 5, that original metadata will still be associated with the row it was first associated with (what is now row 6). If the associated object is deleted its metadata is deleted too.

DeveloperMetadataLocation
A location where metadata may be associated in a spreadsheet.

DeveloperMetadataLookup
Selects DeveloperMetadata that matches all of the specified fields. For example, if only a metadata ID is specified this considers the DeveloperMetadata with that particular unique ID. If a metadata key is specified, this considers all developer metadata with that key. If a key, visibility, and location type are all specified, this considers all developer metadata with that key and visibility that are associated with a location of that type. In general, this selects all DeveloperMetadata that matches the intersection of all the specified fields; any field or combination of fields may be specified.

DimensionGroup
A group over an interval of rows or columns on a sheet, which can contain or be contained within other groups. A group can be collapsed or expanded as a unit on the sheet.

DimensionProperties
Properties about a dimension.

DimensionRange
A range along a single dimension on a sheet. All indexes are zero-based. Indexes are half open: the start index is inclusive and the end index is exclusive. Missing indexes indicate the range is unbounded on that side.

DuplicateFilterViewRequest
Duplicates a particular filter view.

DuplicateFilterViewResponse
The result of a filter view being duplicated.

DuplicateSheetRequest
Duplicates the contents of a sheet.

DuplicateSheetResponse
The result of duplicating a sheet.

Editors
The editors of a protected range.

EmbeddedChart
A chart embedded in a sheet.

EmbeddedObjectBorder
A border along an embedded object.

EmbeddedObjectPosition
The position of an embedded object such as a chart.

ErrorValue
An error in a cell.

ExtendedValue
The kinds of value that a cell in a spreadsheet can have.

FilterCriteria
Criteria for showing/hiding rows in a filter or filter view.

FilterSpec
The filter criteria associated with a specific column.

FilterView
A filter view.

FindReplaceRequest
Finds and replaces data in cells over a range, sheet, or all sheets.

FindReplaceResponse
The result of the find/replace.

GetSpreadsheetByDataFilterRequest
The request for retrieving a Spreadsheet.

GradientRule
A rule that applies a gradient color scale format, based on the interpolation points listed. The format of a cell will vary based on its contents as compared to the values of the interpolation points.

GridCoordinate
A coordinate in a sheet. All indexes are zero-based.

GridData
Data in the grid, as well as metadata about the dimensions.

GridProperties
Properties of a grid.

GridRange
A range on a sheet. All indexes are zero-based. Indexes are half open, i.e. the start index is inclusive and the end index is exclusive -- [start_index, end_index). Missing indexes indicate the range is unbounded on that side. For example, if "Sheet1" is sheet ID 0, then: Sheet1!A1:A1 == sheet_id: 0, start_row_index: 0, end_row_index: 1, start_column_index: 0, end_column_index: 1 Sheet1!A3:B4 == sheet_id: 0, start_row_index: 2, end_row_index: 4, start_column_index: 0, end_column_index: 2 Sheet1!A:B == sheet_id: 0, start_column_index: 0, end_column_index: 2 Sheet1!A5:B == sheet_id: 0, start_row_index: 4, start_column_index: 0, end_column_index: 2 Sheet1 == sheet_id:0 The start index must always be less than or equal to the end index. If the start index equals the end index, then the range is empty. Empty ranges are typically not meaningful and are usually rendered in the UI as #REF!.

HistogramChartSpec
A histogram chart. A histogram chart groups data items into bins, displaying each bin as a column of stacked items. Histograms are used to display the distribution of a dataset. Each column of items represents a range into which those items fall. The number of bins can be chosen automatically or specified explicitly.

HistogramRule
Allows you to organize the numeric values in a source data column into buckets of a constant size. All values from HistogramRule.start to HistogramRule.end are placed into groups of size HistogramRule.interval. In addition, all values below HistogramRule.start are placed in one group, and all values above HistogramRule.end are placed in another. Only HistogramRule.interval is required, though if HistogramRule.start and HistogramRule.end are both provided, HistogramRule.start must be less than HistogramRule.end. For example, a pivot table showing average purchase amount by age that has 50+ rows: +-----+-------------------+ | Age | AVERAGE of Amount | +-----+-------------------+ | 16 | $27.13 | | 17 | $5.24 | | 18 | $20.15 | ... +-----+-------------------+ could be turned into a pivot table that looks like the one below by applying a histogram group rule with a HistogramRule.start of 25, an HistogramRule.interval of 20, and an HistogramRule.end of 65. +-------------+-------------------+ | Grouped Age | AVERAGE of Amount | +-------------+-------------------+ | < 25 | $19.34 | | 25-45 | $31.43 | | 45-65 | $35.87 | | > 65 | $27.55 | +-------------+-------------------+ | Grand Total | $29.12 | +-------------+-------------------+

HistogramSeries
A histogram series containing the series color and data.

InsertDimensionRequest
Inserts rows or columns in a sheet at a particular index.

InsertRangeRequest
Inserts cells into a range, shifting the existing cells over or down.

InterpolationPoint
A single interpolation point on a gradient conditional format. These pin the gradient color scale according to the color, type and value chosen.

Interval
Represents a time interval, encoded as a Timestamp start (inclusive) and a Timestamp end (exclusive). The start must be less than or equal to the end. When the start equals the end, the interval is empty (matches no time). When both start and end are unspecified, the interval matches any time.

IterativeCalculationSettings
Settings to control how circular dependencies are resolved with iterative calculation.

KeyValueFormat
Formatting options for key value.

LineStyle
Properties that describe the style of a line.

Link
An external or local reference.

ManualRule
Allows you to manually organize the values in a source data column into buckets with names of your choosing. For example, a pivot table that aggregates population by state: +-------+-------------------+ | State | SUM of Population | +-------+-------------------+ | AK | 0.7 | | AL | 4.8 | | AR | 2.9 | ... +-------+-------------------+ could be turned into a pivot table that aggregates population by time zone by providing a list of groups (for example, groupName = 'Central', items = ['AL', 'AR', 'IA', ...]) to a manual group rule. Note that a similar effect could be achieved by adding a time zone column to the source data and adjusting the pivot table. +-----------+-------------------+ | Time Zone | SUM of Population | +-----------+-------------------+ | Central | 106.3 | | Eastern | 151.9 | | Mountain | 17.4 | ... +-----------+-------------------+

ManualRuleGroup
A group name and a list of items from the source data that should be placed in the group with this name.

MatchedDeveloperMetadata
A developer metadata entry and the data filters specified in the original request that matched it.

MatchedValueRange
A value range that was matched by one or more data filers.

MergeCellsRequest
Merges all cells in the range.

MoveDimensionRequest
Moves one or more rows or columns.

NamedRange
A named range.

NumberFormat
The number format of a cell.

OrgChartSpec
An org chart. Org charts require a unique set of labels in labels and may optionally include parent_labels and tooltips. parent_labels contain, for each node, the label identifying the parent node. tooltips contain, for each node, an optional tooltip. For example, to describe an OrgChart with Alice as the CEO, Bob as the President (reporting to Alice) and Cathy as VP of Sales (also reporting to Alice), have labels contain "Alice", "Bob", "Cathy", parent_labels contain "", "Alice", "Alice" and tooltips contain "CEO", "President", "VP Sales".

OverlayPosition
The location an object is overlaid on top of a grid.

Padding
The amount of padding around the cell, in pixels. When updating padding, every field must be specified.

PasteDataRequest
Inserts data into the spreadsheet starting at the specified coordinate.

PieChartSpec
A pie chart.

PivotFilterCriteria
Criteria for showing/hiding rows in a pivot table.

PivotFilterSpec
The pivot table filter criteria associated with a specific source column offset.

PivotGroup
A single grouping (either row or column) in a pivot table.

PivotGroupLimit
The count limit on rows or columns in the pivot group.

PivotGroupRule
An optional setting on a PivotGroup that defines buckets for the values in the source data column rather than breaking out each individual value. Only one PivotGroup with a group rule may be added for each column in the source data, though on any given column you may add both a PivotGroup that has a rule and a PivotGroup that does not.

PivotGroupSortValueBucket
Information about which values in a pivot group should be used for sorting.

PivotGroupValueMetadata
Metadata about a value in a pivot grouping.

PivotTable
A pivot table.

PivotValue
The definition of how a value in a pivot table should be calculated.

PointStyle
The style of a point on the chart.

ProtectedRange
A protected range.

RandomizeRangeRequest
Randomizes the order of the rows in a range.

RefreshDataSourceObjectExecutionStatus
The execution status of refreshing one data source object.

RefreshDataSourceRequest
Refreshes one or multiple data source objects in the spreadsheet by the specified references. The request requires an additional bigquery.readonly OAuth scope. If there are multiple refresh requests referencing the same data source objects in one batch, only the last refresh request is processed, and all those requests will have the same response accordingly.

RefreshDataSourceResponse
The response from refreshing one or multiple data source objects.

RepeatCellRequest
Updates all cells in the range to the values in the given Cell object. Only the fields listed in the fields field are updated; others are unchanged. If writing a cell with a formula, the formula's ranges will automatically increment for each field in the range. For example, if writing a cell with formula =A1 into range B2:C4, B2 would be =A1, B3 would be =A2, B4 would be =A3, C2 would be =B1, C3 would be =B2, C4 would be =B3. To keep the formula's ranges static, use the $ indicator. For example, use the formula =$A$1 to prevent both the row and the column from incrementing.

Request
A single kind of update to apply to a spreadsheet.

Response
A single response from an update.

RowData
Data about each cell in a row.

ScorecardChartSpec
A scorecard chart. Scorecard charts are used to highlight key performance indicators, known as KPIs, on the spreadsheet. A scorecard chart can represent things like total sales, average cost, or a top selling item. You can specify a single data value, or aggregate over a range of data. Percentage or absolute difference from a baseline value can be highlighted, like changes over time.

SearchDeveloperMetadataRequest
A request to retrieve all developer metadata matching the set of specified criteria.

SearchDeveloperMetadataResponse
A reply to a developer metadata search request.

SetBasicFilterRequest
Sets the basic filter associated with a sheet.

SetDataValidationRequest
Sets a data validation rule to every cell in the range. To clear validation in a range, call this with no rule specified.

Sheet
A sheet in a spreadsheet.

SheetProperties
Properties of a sheet.

Sheets
Central instance to access all Sheets related resource activities

Slicer
A slicer in a sheet.

SlicerSpec
The specifications of a slicer.

SortRangeRequest
Sorts data in rows based on a sort order per column.

SortSpec
A sort order associated with a specific column or row.

SourceAndDestination
A combination of a source range and how to extend that source.

Spreadsheet
Resource that represents a spreadsheet.

SpreadsheetBatchUpdateCall
Applies one or more updates to the spreadsheet. Each request is validated before being applied. If any request is not valid then the entire request will fail and nothing will be applied. Some requests have replies to give you some information about how they are applied. The replies will mirror the requests. For example, if you applied 4 updates and the 3rd one had a reply, then the response will have 2 empty replies, the actual reply, and another empty reply, in that order. Due to the collaborative nature of spreadsheets, it is not guaranteed that the spreadsheet will reflect exactly your changes after this completes, however it is guaranteed that the updates in the request will be applied together atomically. Your changes may be altered with respect to collaborator changes. If there are no collaborators, the spreadsheet should reflect your changes.

SpreadsheetCreateCall
Creates a spreadsheet, returning the newly created spreadsheet.

SpreadsheetDeveloperMetadataGetCall
Returns the developer metadata with the specified ID. The caller must specify the spreadsheet ID and the developer metadata's unique metadataId.

SpreadsheetDeveloperMetadataSearchCall
Returns all developer metadata matching the specified DataFilter. If the provided DataFilter represents a DeveloperMetadataLookup object, this will return all DeveloperMetadata entries selected by it. If the DataFilter represents a location in a spreadsheet, this will return all developer metadata associated with locations intersecting that region.

SpreadsheetGetByDataFilterCall
Returns the spreadsheet at the given ID. The caller must specify the spreadsheet ID. This method differs from GetSpreadsheet in that it allows selecting which subsets of spreadsheet data to return by specifying a dataFilters parameter. Multiple DataFilters can be specified. Specifying one or more data filters will return the portions of the spreadsheet that intersect ranges matched by any of the filters. By default, data within grids will not be returned. You can include grid data one of two ways: * Specify a field mask listing your desired fields using the fields URL parameter in HTTP * Set the includeGridData parameter to true. If a field mask is set, the includeGridData parameter is ignored For large spreadsheets, it is recommended to retrieve only the specific fields of the spreadsheet that you want.

SpreadsheetGetCall
Returns the spreadsheet at the given ID. The caller must specify the spreadsheet ID. By default, data within grids will not be returned. You can include grid data one of two ways: * Specify a field mask listing your desired fields using the fields URL parameter in HTTP * Set the includeGridData URL parameter to true. If a field mask is set, the includeGridData parameter is ignored For large spreadsheets, it is recommended to retrieve only the specific fields of the spreadsheet that you want. To retrieve only subsets of the spreadsheet, use the ranges URL parameter. Multiple ranges can be specified. Limiting the range will return only the portions of the spreadsheet that intersect the requested ranges. Ranges are specified using A1 notation.

SpreadsheetMethods
A builder providing access to all methods supported on spreadsheet resources. It is not used directly, but through the Sheets hub.

SpreadsheetProperties
Properties of a spreadsheet.

SpreadsheetSheetCopyToCall
Copies a single sheet from a spreadsheet to another spreadsheet. Returns the properties of the newly created sheet.

SpreadsheetTheme
Represents spreadsheet theme

SpreadsheetValueAppendCall
Appends values to a spreadsheet. The input range is used to search for existing data and find a "table" within that range. Values will be appended to the next row of the table, starting with the first column of the table. See the guide and sample code for specific details of how tables are detected and data is appended. The caller must specify the spreadsheet ID, range, and a valueInputOption. The valueInputOption only controls how the input data will be added to the sheet (column-wise or row-wise), it does not influence what cell the data starts being written to.

SpreadsheetValueBatchClearByDataFilterCall
Clears one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more DataFilters. Ranges matching any of the specified data filters will be cleared. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueBatchClearCall
Clears one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more ranges. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueBatchGetByDataFilterCall
Returns one or more ranges of values that match the specified data filters. The caller must specify the spreadsheet ID and one or more DataFilters. Ranges that match any of the data filters in the request will be returned.

SpreadsheetValueBatchGetCall
Returns one or more ranges of values from a spreadsheet. The caller must specify the spreadsheet ID and one or more ranges.

SpreadsheetValueBatchUpdateByDataFilterCall
Sets values in one or more ranges of a spreadsheet. The caller must specify the spreadsheet ID, a valueInputOption, and one or more DataFilterValueRanges.

SpreadsheetValueBatchUpdateCall
Sets values in one or more ranges of a spreadsheet. The caller must specify the spreadsheet ID, a valueInputOption, and one or more ValueRanges.

SpreadsheetValueClearCall
Clears values from a spreadsheet. The caller must specify the spreadsheet ID and range. Only values are cleared -- all other properties of the cell (such as formatting, data validation, etc..) are kept.

SpreadsheetValueGetCall
Returns a range of values from a spreadsheet. The caller must specify the spreadsheet ID and a range.

SpreadsheetValueUpdateCall
Sets values in a range of a spreadsheet. The caller must specify the spreadsheet ID, range, and a valueInputOption.

TextFormat
The format of a run of text in a cell. Absent values indicate that the field isn't specified.

TextFormatRun
A run of a text format. The format of this run continues until the start index of the next run. When updating, all fields must be set.

TextPosition
Position settings for text.

TextRotation
The rotation applied to text in a cell.

TextToColumnsRequest
Splits a column of text into multiple columns, based on a delimiter in each cell.

ThemeColorPair
A pair mapping a spreadsheet theme color type to the concrete color it represents.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

TreemapChartColorScale
A color scale for a treemap chart.

TreemapChartSpec
A Treemap chart.

TrimWhitespaceRequest
Trims the whitespace (such as spaces, tabs, or new lines) in every cell in the specified range. This request removes all whitespace from the start and end of each cell's text, and reduces any subsequence of remaining whitespace characters to a single space. If the resulting trimmed text starts with a '+' or '=' character, the text remains as a string value and isn't interpreted as a formula.

TrimWhitespaceResponse
The result of trimming whitespace in cells.

UnmergeCellsRequest
Unmerges cells in the given range.

UpdateBandingRequest
Updates properties of the supplied banded range.

UpdateBordersRequest
Updates the borders of a range. If a field is not set in the request, that means the border remains as-is. For example, with two subsequent UpdateBordersRequest: 1. range: A1:A5 { top: RED, bottom: WHITE } 2. range: A1:A5 { left: BLUE } That would result in A1:A5 having a borders of { top: RED, bottom: WHITE, left: BLUE }. If you want to clear a border, explicitly set the style to NONE.

UpdateCellsRequest
Updates all cells in a range with new data.

UpdateChartSpecRequest
Updates a chart's specifications. (This does not move or resize a chart. To move or resize a chart, use UpdateEmbeddedObjectPositionRequest.)

UpdateConditionalFormatRuleRequest
Updates a conditional format rule at the given index, or moves a conditional format rule to another index.

UpdateConditionalFormatRuleResponse
The result of updating a conditional format rule.

UpdateDataSourceRequest
Updates a data source. After the data source is updated successfully, an execution is triggered to refresh the associated DATA_SOURCE sheet to read data from the updated data source. The request requires an additional bigquery.readonly OAuth scope.

UpdateDataSourceResponse
The response from updating data source.

UpdateDeveloperMetadataRequest
A request to update properties of developer metadata. Updates the properties of the developer metadata selected by the filters to the values provided in the DeveloperMetadata resource. Callers must specify the properties they wish to update in the fields parameter, as well as specify at least one DataFilter matching the metadata they wish to update.

UpdateDeveloperMetadataResponse
The response from updating developer metadata.

UpdateDimensionGroupRequest
Updates the state of the specified group.

UpdateDimensionPropertiesRequest
Updates properties of dimensions within the specified range.

UpdateEmbeddedObjectBorderRequest
Updates an embedded object's border property.

UpdateEmbeddedObjectPositionRequest
Update an embedded object's position (such as a moving or resizing a chart or image).

UpdateEmbeddedObjectPositionResponse
The result of updating an embedded object's position.

UpdateFilterViewRequest
Updates properties of the filter view.

UpdateNamedRangeRequest
Updates properties of the named range with the specified namedRangeId.

UpdateProtectedRangeRequest
Updates an existing protected range with the specified protectedRangeId.

UpdateSheetPropertiesRequest
Updates properties of the sheet with the specified sheetId.

UpdateSlicerSpecRequest
Updates a slicer's specifications. (This does not move or resize a slicer. To move or resize a slicer use UpdateEmbeddedObjectPositionRequest.

UpdateSpreadsheetPropertiesRequest
Updates properties of a spreadsheet.

UpdateValuesByDataFilterResponse
The response when updating a range of values by a data filter in a spreadsheet.

UpdateValuesResponse
The response when updating a range of values in a spreadsheet.

ValueRange
Data within a range of the spreadsheet.

WaterfallChartColumnStyle
Styles for a waterfall chart column.

WaterfallChartCustomSubtotal
A custom subtotal column for a waterfall chart series.

WaterfallChartDomain
The domain of a waterfall chart.

WaterfallChartSeries
A single series of data for a waterfall chart.

WaterfallChartSpec
A waterfall chart.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.