Struct google_sheets4::api::SheetProperties

pub struct SheetProperties {
    pub data_source_sheet_properties: Option<DataSourceSheetProperties>,
    pub grid_properties: Option<GridProperties>,
    pub hidden: Option<bool>,
    pub index: Option<i32>,
    pub right_to_left: Option<bool>,
    pub sheet_id: Option<i32>,
    pub sheet_type: Option<String>,
    pub tab_color: Option<Color>,
    pub tab_color_style: Option<ColorStyle>,
    pub title: Option<String>,
}

Properties of a sheet.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

sheets copy to spreadsheets (response)

Fields
data_source_sheet_properties: Option<DataSourceSheetProperties>

Output only. If present, the field contains DATA_SOURCE sheet specific properties.

grid_properties: Option<GridProperties>

Additional properties of the sheet if this sheet is a grid. (If the sheet is an object sheet, containing a chart or image, then this field will be absent.) When writing it is an error to set any grid properties on non-grid sheets. If this sheet is a DATA_SOURCE sheet, this field is output only but contains the properties that reflect how a data source sheet is rendered in the UI, e.g. row_count.

hidden: Option<bool>

True if the sheet is hidden in the UI, false if it's visible.

index: Option<i32>

The index of the sheet within the spreadsheet. When adding or updating sheet properties, if this field is excluded then the sheet is added or moved to the end of the sheet list. When updating sheet indices or inserting sheets, movement is considered in "before the move" indexes. For example, if there were 3 sheets (S1, S2, S3) in order to move S1 ahead of S2 the index would have to be set to 2. A sheet index update request is ignored if the requested index is identical to the sheets current index or if the requested new index is equal to the current sheet index + 1.

right_to_left: Option<bool>

True if the sheet is an RTL sheet instead of an LTR sheet.

sheet_id: Option<i32>

The ID of the sheet. Must be non-negative. This field cannot be changed once set.

sheet_type: Option<String>

The type of sheet. Defaults to GRID. This field cannot be changed once set.

tab_color: Option<Color>

The color of the tab in the UI.

tab_color_style: Option<ColorStyle>

The color of the tab in the UI. If tab_color is also set, this field takes precedence.

title: Option<String>

The name of the sheet.