Struct google_sheets4::Request

pub struct Request {
    pub duplicate_filter_view: Option<DuplicateFilterViewRequest>,
    pub sort_range: Option<SortRangeRequest>,
    pub update_embedded_object_position: Option<UpdateEmbeddedObjectPositionRequest>,
    pub update_conditional_format_rule: Option<UpdateConditionalFormatRuleRequest>,
    pub update_dimension_group: Option<UpdateDimensionGroupRequest>,
    pub delete_dimension: Option<DeleteDimensionRequest>,
    pub cut_paste: Option<CutPasteRequest>,
    pub delete_embedded_object: Option<DeleteEmbeddedObjectRequest>,
    pub paste_data: Option<PasteDataRequest>,
    pub append_cells: Option<AppendCellsRequest>,
    pub duplicate_sheet: Option<DuplicateSheetRequest>,
    pub update_sheet_properties: Option<UpdateSheetPropertiesRequest>,
    pub add_chart: Option<AddChartRequest>,
    pub add_conditional_format_rule: Option<AddConditionalFormatRuleRequest>,
    pub delete_banding: Option<DeleteBandingRequest>,
    pub repeat_cell: Option<RepeatCellRequest>,
    pub find_replace: Option<FindReplaceRequest>,
    pub auto_fill: Option<AutoFillRequest>,
    pub set_basic_filter: Option<SetBasicFilterRequest>,
    pub update_spreadsheet_properties: Option<UpdateSpreadsheetPropertiesRequest>,
    pub update_cells: Option<UpdateCellsRequest>,
    pub randomize_range: Option<RandomizeRangeRequest>,
    pub append_dimension: Option<AppendDimensionRequest>,
    pub update_banding: Option<UpdateBandingRequest>,
    pub add_banding: Option<AddBandingRequest>,
    pub auto_resize_dimensions: Option<AutoResizeDimensionsRequest>,
    pub update_dimension_properties: Option<UpdateDimensionPropertiesRequest>,
    pub text_to_columns: Option<TextToColumnsRequest>,
    pub unmerge_cells: Option<UnmergeCellsRequest>,
    pub set_data_validation: Option<SetDataValidationRequest>,
    pub delete_range: Option<DeleteRangeRequest>,
    pub clear_basic_filter: Option<ClearBasicFilterRequest>,
    pub add_named_range: Option<AddNamedRangeRequest>,
    pub update_chart_spec: Option<UpdateChartSpecRequest>,
    pub insert_range: Option<InsertRangeRequest>,
    pub update_developer_metadata: Option<UpdateDeveloperMetadataRequest>,
    pub move_dimension: Option<MoveDimensionRequest>,
    pub update_protected_range: Option<UpdateProtectedRangeRequest>,
    pub create_developer_metadata: Option<CreateDeveloperMetadataRequest>,
    pub delete_dimension_group: Option<DeleteDimensionGroupRequest>,
    pub merge_cells: Option<MergeCellsRequest>,
    pub delete_protected_range: Option<DeleteProtectedRangeRequest>,
    pub add_filter_view: Option<AddFilterViewRequest>,
    pub delete_developer_metadata: Option<DeleteDeveloperMetadataRequest>,
    pub delete_sheet: Option<DeleteSheetRequest>,
    pub update_borders: Option<UpdateBordersRequest>,
    pub add_protected_range: Option<AddProtectedRangeRequest>,
    pub copy_paste: Option<CopyPasteRequest>,
    pub add_sheet: Option<AddSheetRequest>,
    pub delete_named_range: Option<DeleteNamedRangeRequest>,
    pub delete_filter_view: Option<DeleteFilterViewRequest>,
    pub update_named_range: Option<UpdateNamedRangeRequest>,
    pub insert_dimension: Option<InsertDimensionRequest>,
    pub update_filter_view: Option<UpdateFilterViewRequest>,
    pub delete_conditional_format_rule: Option<DeleteConditionalFormatRuleRequest>,
    pub add_dimension_group: Option<AddDimensionGroupRequest>,
}

A single kind of update to apply to a spreadsheet.

This type is not used in any activity, and only used as part of another schema.

Fields
duplicate_filter_view: Option<DuplicateFilterViewRequest>

Duplicates a filter view.

sort_range: Option<SortRangeRequest>

Sorts data in a range.

update_embedded_object_position: Option<UpdateEmbeddedObjectPositionRequest>

Updates an embedded object's (e.g. chart, image) position.

update_conditional_format_rule: Option<UpdateConditionalFormatRuleRequest>

Updates an existing conditional format rule.

update_dimension_group: Option<UpdateDimensionGroupRequest>

Updates the state of the specified group.

delete_dimension: Option<DeleteDimensionRequest>

Deletes rows or columns in a sheet.

cut_paste: Option<CutPasteRequest>

Cuts data from one area and pastes it to another.

delete_embedded_object: Option<DeleteEmbeddedObjectRequest>

Deletes an embedded object (e.g, chart, image) in a sheet.

paste_data: Option<PasteDataRequest>

Pastes data (HTML or delimited) into a sheet.

append_cells: Option<AppendCellsRequest>

Appends cells after the last row with data in a sheet.

duplicate_sheet: Option<DuplicateSheetRequest>

Duplicates a sheet.

update_sheet_properties: Option<UpdateSheetPropertiesRequest>

Updates a sheet's properties.

add_chart: Option<AddChartRequest>

Adds a chart.

add_conditional_format_rule: Option<AddConditionalFormatRuleRequest>

Adds a new conditional format rule.

delete_banding: Option<DeleteBandingRequest>

Removes a banded range

repeat_cell: Option<RepeatCellRequest>

Repeats a single cell across a range.

find_replace: Option<FindReplaceRequest>

Finds and replaces occurrences of some text with other text.

auto_fill: Option<AutoFillRequest>

Automatically fills in more data based on existing data.

set_basic_filter: Option<SetBasicFilterRequest>

Sets the basic filter on a sheet.

update_spreadsheet_properties: Option<UpdateSpreadsheetPropertiesRequest>

Updates the spreadsheet's properties.

update_cells: Option<UpdateCellsRequest>

Updates many cells at once.

randomize_range: Option<RandomizeRangeRequest>

Randomizes the order of the rows in a range.

append_dimension: Option<AppendDimensionRequest>

Appends dimensions to the end of a sheet.

update_banding: Option<UpdateBandingRequest>

Updates a banded range

add_banding: Option<AddBandingRequest>

Adds a new banded range

auto_resize_dimensions: Option<AutoResizeDimensionsRequest>

Automatically resizes one or more dimensions based on the contents of the cells in that dimension.

update_dimension_properties: Option<UpdateDimensionPropertiesRequest>

Updates dimensions' properties.

text_to_columns: Option<TextToColumnsRequest>

Converts a column of text into many columns of text.

unmerge_cells: Option<UnmergeCellsRequest>

Unmerges merged cells.

set_data_validation: Option<SetDataValidationRequest>

Sets data validation for one or more cells.

delete_range: Option<DeleteRangeRequest>

Deletes a range of cells from a sheet, shifting the remaining cells.

clear_basic_filter: Option<ClearBasicFilterRequest>

Clears the basic filter on a sheet.

add_named_range: Option<AddNamedRangeRequest>

Adds a named range.

update_chart_spec: Option<UpdateChartSpecRequest>

Updates a chart's specifications.

insert_range: Option<InsertRangeRequest>

Inserts new cells in a sheet, shifting the existing cells.

update_developer_metadata: Option<UpdateDeveloperMetadataRequest>

Updates an existing developer metadata entry

move_dimension: Option<MoveDimensionRequest>

Moves rows or columns to another location in a sheet.

update_protected_range: Option<UpdateProtectedRangeRequest>

Updates a protected range.

create_developer_metadata: Option<CreateDeveloperMetadataRequest>

Creates new developer metadata

delete_dimension_group: Option<DeleteDimensionGroupRequest>

Deletes a group over the specified range.

merge_cells: Option<MergeCellsRequest>

Merges cells together.

delete_protected_range: Option<DeleteProtectedRangeRequest>

Deletes a protected range.

add_filter_view: Option<AddFilterViewRequest>

Adds a filter view.

delete_developer_metadata: Option<DeleteDeveloperMetadataRequest>

Deletes developer metadata

delete_sheet: Option<DeleteSheetRequest>

Deletes a sheet.

update_borders: Option<UpdateBordersRequest>

Updates the borders in a range of cells.

add_protected_range: Option<AddProtectedRangeRequest>

Adds a protected range.

copy_paste: Option<CopyPasteRequest>

Copies data from one area and pastes it to another.

add_sheet: Option<AddSheetRequest>

Adds a sheet.

delete_named_range: Option<DeleteNamedRangeRequest>

Deletes a named range.

delete_filter_view: Option<DeleteFilterViewRequest>

Deletes a filter view from a sheet.

update_named_range: Option<UpdateNamedRangeRequest>

Updates a named range.

insert_dimension: Option<InsertDimensionRequest>

Inserts new rows or columns in a sheet.

update_filter_view: Option<UpdateFilterViewRequest>

Updates the properties of a filter view.

delete_conditional_format_rule: Option<DeleteConditionalFormatRuleRequest>

Deletes an existing conditional format rule.

add_dimension_group: Option<AddDimensionGroupRequest>

Creates a group over the specified range.